src/Gastromio.Core/Application/Commands/ChangePassword/ChangePasswordCommandHandler.cs
using System;
using System.Threading;
using System.Threading.Tasks;
using Gastromio.Core.Application.Ports.Persistence;
using Gastromio.Core.Common;
using Gastromio.Core.Domain.Model.Users;

namespace Gastromio.Core.Application.Commands.ChangePassword
{
    public class ChangePasswordCommandHandler : ICommandHandler<ChangePasswordCommand, bool>
    {
        private readonly IUserRepository userRepository;

        public ChangePasswordCommandHandler(IUserRepository userRepository)
        {
            this.userRepository = userRepository;
        }

        public async Task<Result<bool>> HandleAsync(ChangePasswordCommand command, User currentUser, CancellationToken cancellationToken = default)
        {
            if (command == null)
                throw new ArgumentNullException(nameof(command));

            if (currentUser == null)
                return FailureResult<bool>.Unauthorized();           

            var result = currentUser.ChangePassword(command.Password, true, currentUser.Id);
            if (result.IsFailure)
                return result;

            await userRepository.StoreAsync(currentUser, cancellationToken);

            return SuccessResult<bool>.Create(true);
        }
    }
}
